fix: lint 里翻出两个真 bug——跨时区差一天、必崩的 NameError
430 → 360，但重点不是数字，是这批里有两个真问题。

## DTZ011：证书有效期判定每天有 8 小时是错的

实测：宿主机 CST +0800，**容器 UTC +0000**。5 处裸用 date.today() 拿到的是
UTC 日期，于是每天 00:00–08:00 这段时间它比业务日期少一天。

落到业务上：一张 2026-08-14 到期的焊工证，在 08-15 凌晨会被判成仍然有效——
有效期判定整整差一天。同类还有规则版本切换日（review_date >= 2026-08-01）
和标准现行性核验。

仓库里本来就有正确的口径（server_time 用 SERVER_TZ，默认 Asia/Shanghai），
这 5 处只是各写各的绕过了它。新增 business_today() 与之同源，5 处全部改用。
测试构造 UTC 与业务日期真正不同的时刻（UTC 08-14 17:30 = CST 08-15 01:30）
钉住差异，而不是只断言「两个函数返回一样」——那种断言在任何时区都通过。

## F821：CheckResult 从未定义，而调用路径是活的

scripts/deployment_report.py 用了 CheckResult 四次，文件里没有定义、也没有
import。release_runtime_check 在 879 行被真实调用，只要传了 --release-manifest
就是 NameError。类就在 scripts/verify_deployment.py 里，同一行 import 漏了它。

## 其余

SIM103/C401/F841/RUF046/RUF059/PIE810/FURB192 等 64 条机械修复；
10 个带 shebang 的脚本补上可执行位。全量 1681 个测试通过。

+++ backend/apps/ocr_service/pages.py
@@ -140,7 +140,7 @@ def render_pdf_pages(
                     page_path,
                     page_no=page_no,
                     source_type="pdf",
-                    render_dpi=int(round(scale * 72)),
+                    render_dpi=round(scale * 72),
                     document_path=source_path,
                     rotation=int(page.rotation or 0),
                     source_width=float(page.rect.width),
@@ -154,7 +154,7 @@ def render_pdf_pages(
                     pdf_pixmap_y=int(getattr(pixmap, "y", 0) or 0),
                     page_render_version=PAGE_RENDER_VERSION,
                     requested_render_dpi=dpi,
-                    effective_render_dpi=int(round(scale * 72)),
+                    effective_render_dpi=round(scale * 72),
                     total_pages=total_pages,
                     rendered_pages=rendered_pages,
                     truncated=truncated,
@@ -430,7 +430,7 @@ def render_pdf_page_preview(
             target,
             page_no=page_no,
             source_type="pdf",
-            render_dpi=int(round(scale * 72)),
+            render_dpi=round(scale * 72),
             document_path=source_path,
             rotation=int(page.rotation or 0),
             source_width=float(page.rect.width),
@@ -444,7 +444,7 @@ def render_pdf_page_preview(
             pdf_pixmap_y=int(getattr(pixmap, "y", 0) or 0),
             page_render_version=PAGE_RENDER_VERSION,
             requested_render_dpi=dpi,
-            effective_render_dpi=int(round(scale * 72)),
+            effective_render_dpi=round(scale * 72),
             total_pages=total_pages,
             rendered_pages=[page_no],
             truncated=total_pages > 1,
@@ -588,7 +588,7 @@ def constrain_image_page(
     if target.exists():
         return target
     ratio = max_long_side / float(max(width, height))
-    new_size = (max(1, int(round(width * ratio))), max(1, int(round(height * ratio))))
+    new_size = (max(1, round(width * ratio)), max(1, round(height * ratio)))
     if resize_with_pillow(path, target, new_size) or resize_with_cv2(path, target, new_size):
         return target
     return path
